In the server folder index2.js has overwritten index.js. Therefore index2.js has become index.js.
